requirements.txt
torch==2.8.0+cu128
transformers==4.56.2
trl>=0.12.0
peft>=0.13.0
bitsandbytes>=0.44.0
datasets>=3.0.0
accelerate>=1.0.0
evalplus>=0.3.0
pyyaml>=6.0
jsonlines>=4.0
numpy>=1.26.0
tqdm>=4.66.0
wandb>=0.18.0
